moved test fixtures to coftest.py
